Added the extra ManifestThumbnails Provider
