+++ packages/canvas-panel-core/src/manifesto/ManifestThumbnails/ManifestThumbnails.md
@@ -1,0 +1,1 @@
+Returns all thumbnails from the manifest.
